CODE/LEGACY_ROOT/requirements.txt
langchain
openai
azure-identity
python-dotenv
requests
langchain-openai
pyodbc
pyadomd
msal
